Remove graphics of coordinate comparison
OpenGL coordinates has the origin at the center at the screen and y-axis points up. Screen coordinate on the other hand, has the origin at the top left of the screen and the y-axis points down. Thus, the mapping value should be (200, 150) whereas on the source text, the mapping value is (200, 450). I'm sure the source is incorrect.

The coordinate comparison will be back on chapter 1.2_hello_triangle.