Add task queue for general task and window task
The queue for window task is using the previous one renamed from enqueueTask() to enqueueWindowTask() for adding to the queue. The former name is then used for the general taks. For general task, user does not need to pass a window id (since it does not makes sense). If a user enqueue a task related to window in a general task queue, it may lead to unwanted consequences (data race possibly but not limited to just that).