Remove OpenGL calls from Window class and move them to the user of it
This change is done to make sure the separation of concern is clear. The Window class supposed to wrap GLFWwindow and make it easier to use. On the other hand OpenGL calls is not the concern of this window class and it should not do any OpenGL calls. GLFW function calls are allowed in the class.